+++ packages/code-du-travail-frontend/src/modules/common/Autocomplete/index.ts
@@ -1,2 +1,2 @@
 export * from "./Autocomplete";
-export * from "./AutocompleteV2";
+export * from "./HomemadeAutocomplete";
