+++ packages/code-du-travail-frontend/src/modules/recherche/modal/MinSearchLengthHint.tsx
@@ -5,12 +5,14 @@ interface MinSearchLengthHintProps {
   isVisible: boolean;
   minSearchLength: number;
   variant?: "mobile" | "desktop";
+  paragraphId: string;
 }
 
 export const MinSearchLengthHint = ({
   isVisible,
   minSearchLength,
   variant = "desktop",
+  paragraphId,
 }: MinSearchLengthHintProps) => {
   if (!isVisible) return null;
 
@@ -22,7 +24,7 @@ export const MinSearchLengthHint = ({
           : fr.cx("fr-hidden", "fr-unhidden-md")
       }
     >
-      <p className={hintTextStyle}>
+      <p id={paragraphId} className={hintTextStyle}>
         Tapez {minSearchLength} caractères ou plus pour lancer une recherche
       </p>
     </div>
